requirements.txt
Django==5.0.1
djangorestframework==3.14.0
Pillow==10.2.0
python-dotenv==1.0.0
drf-yasg==1.21.7
django-filter==23.5
django-cors-headers==3.14.0
djangorestframework-simplejwt==5.3.1